// src/types/asaas.ts
export type PaymentAsaas = {
  object: string;
  id: string;
  dateCreated: string;
  customer: string;
  subscription: string;
  paymentLink: null;
  value: number;
  netValue: number;
  originalValue: null;
  interestValue: null;
  description: string;
  billingType: string;
  confirmedDate: null;
  creditCard: {
    creditCardNumber: string;
    creditCardBrand: string;
    creditCardToken: string;
  };
  pixTransaction: null;
  status: string;
  dueDate: string;
  originalDueDate: string;
  paymentDate: null;
  clientPaymentDate: null;
  installmentNumber: null;
  invoiceUrl: string;
  invoiceNumber: string;
  externalReference: null;
  deleted: boolean;
  anticipated: boolean;
  anticipable: boolean;
  creditDate: null;
  estimatedCreditDate: null;
  transactionReceiptUrl: null;
  nossoNumero: null;
  bankSlipUrl: null;
  lastInvoiceViewedDate: null;
  lastBankSlipViewedDate: null;
  discount: {
    value: number;
    limitDate: null;
    dueDateLimitDays: 0;
    type: string;
  };
  fine: { value: number; type: string };
  interest: { value: number; type: string };
  postalService: boolean;
  custody: null;
  refunds: null;
};